// app/src/generated/models/Sh_roomsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Sh_roomssh_status = {
  1: 'Available',
  2: 'Occupied',
  3: 'NeedsReset'
} as const;
export type Sh_roomssh_status = keyof typeof Sh_roomssh_status;
export const Sh_roomsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Sh_roomsstatecode = keyof typeof Sh_roomsstatecode;
export const Sh_roomsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Sh_roomsstatuscode = keyof typeof Sh_roomsstatuscode;

export interface Sh_roomsBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  sh_eventid?: string;
  sh_lastreset?: string;
  sh_lastshare?: string;
  "sh_LocationId@odata.bind"?: string;
  sh_meetinglink?: string;
  sh_name?: string;
  sh_patientname?: string;
  sh_roomid: string;
  sh_roomupn?: string;
  sh_sharedwith?: number;
  sh_shareexternally?: boolean;
  sh_status?: Sh_roomssh_status;
  "sh_SubLocationId@odata.bind"?: string;
  statecode: Sh_roomsstatecode;
  statuscode?: Sh_roomsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Sh_rooms extends Sh_roomsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  sh_locationidname?: string;
  sh_shareexternallyname?: string;
  sh_statusname?: string;
  sh_sublocationidname?: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  sh_locationid?: object;
  _sh_locationid_value?: string;
  sh_sublocationid?: object;
  _sh_sublocationid_value?: string;
}